requirements.txt
streamlit
PyPDF2
sentence-transformers
endee
google-generativeai
python-dotenv
